pyproject.toml
[project]
name = "beatguessr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.9.0",
    "requests>=2.32.5",
]
